requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-crispy-forms==1.6.1
django-environ==0.4.1
django-rest-auth==0.9.2
django-rest-framework==0.1.0
djangorestframework==3.7.3
drf-writable-nested==0.3.2
idna==2.6
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
